code/Function/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.6.1
decorator==5.1.1
deprecation==2.1.0
et-xmlfile==1.1.0
Flask==2.2.2
frozenlist==1.3.1
fsspec==2022.8.2
functions-framework==3.2.0
gcsfs==2022.8.2
google-api-core==2.10.0
google-auth==2.11.0
google-auth-oauthlib==0.5.2
google-cloud-appengine-logging==1.1.4
google-cloud-audit-log==0.2.4
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.14.2
google-cloud-core==2.3.2
google-cloud-logging==3.2.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.48.1
grpcio-status==1.48.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.2
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.4
proto-plus==1.22.1
protobuf==4.21.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
watchdog==2.1.9
Werkzeug==2.2.2
xlrd==2.0.1
yarl==1.8.1
